+++ src/AlgoTradeForge.Domain/Strategy/IOrderContext.cs
@@ -4,7 +4,7 @@ namespace AlgoTradeForge.Domain.Strategy;
 
 public interface IOrderContext
 {
-    decimal Cash { get; }
+    long Cash { get; }
     long Submit(Order order);
     bool Cancel(long orderId);
     IReadOnlyList<Order> GetPendingOrders();
