+++ src/AlgoTradeForge.Domain/Trading/TakeProfitLevel.cs
@@ -1,3 +1,3 @@
 namespace AlgoTradeForge.Domain.Trading;
 
-public readonly record struct TakeProfitLevel(decimal Price, decimal ClosurePercentage);
+public readonly record struct TakeProfitLevel(long Price, decimal ClosurePercentage);
